fix(extract): address follow-up review feedback
Document extract's command-specific max-pages default and replace direct loopback test mutations with a shared guard that restores the previous thread-local value.

Resolves review thread PRRT_kwDORS2O8s6Gmsm1

Resolves review thread PRRT_kwDORS2O8s6GmtPK

Resolves review thread PRRT_kwDORS2O8s6Gmsm8
